requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==21.2.0

# AWS / Storage
boto3>=1.26,<1.27
django-storages==1.14.2

# Celery / Redis
celery==5.3.4
redis==5.0.1

# Utils
python-dotenv==1.0.0
Pillow==10.1.0

python-decouple

django-filter
